gcp/gae/gae_rec/requirements.txt
Flask==1.1.1
gunicorn==19.9.0
joblib==0.14.0
implicit==0.4.2
google-cloud-storage==1.21.0
uwsgi==2.0.18